requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
crispy-bootstrap4==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-filter==24.3
isort==6.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
sqlparse==0.5.3
